MOSU-40 refactor: 도메인 패키지 이동
