+++ src/main/java/life/mosu/mosuserver/domain/faq/FaqFile.java
@@ -1,6 +1,8 @@
-package life.mosu.mosuserver.infra.storage.domain;
+package life.mosu.mosuserver.domain.faq;
 
 import jakarta.persistence.Embeddable;
+import life.mosu.mosuserver.infra.storage.domain.File;
+import life.mosu.mosuserver.infra.storage.domain.Visibility;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,10 +11,10 @@ import lombok.NoArgsConstructor;
 @Getter
 @Embeddable
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class EventImage extends File {
+public class FaqFile extends File {
 
     @Builder
-    public EventImage(String fileName, String s3Key, Visibility visibility) {
+    public FaqFile(String fileName, String s3Key, Visibility visibility) {
         super(fileName, s3Key, visibility);
     }
 }
