+++ src/main/java/life/mosu/mosuserver/domain/inquiry/InquiryFile.java
@@ -1,6 +1,8 @@
-package life.mosu.mosuserver.infra.storage.domain;
+package life.mosu.mosuserver.domain.inquiry;
 
 import jakarta.persistence.Embeddable;
+import life.mosu.mosuserver.infra.storage.domain.File;
+import life.mosu.mosuserver.infra.storage.domain.Visibility;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,10 +11,10 @@ import lombok.NoArgsConstructor;
 @Getter
 @Embeddable
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class NoticeFile extends File {
+public class InquiryFile extends File {
 
     @Builder
-    public NoticeFile(String fileName, String s3Key, Visibility visibility) {
+    public InquiryFile(String fileName, String s3Key, Visibility visibility) {
         super(fileName, s3Key, visibility);
     }
 }
